// src/app/Interfaces/Productos/producto.model.ts
export interface Producto {
    codigo: number | null;
    codigo2: string | null;
    nombreUnico: string | null;
    fotourl: string | null;
    descripcion: string | null;
    existenciaMinima: number | null;
    existenciaMaxima: number | null;
    stockactual: number | null;
    costopromediokardex: number | null;
    costoultimokardex: number | null;
    puedeVenderse: number | null;
    puedeComprarse: number | null;
    deAlquiler: number | null;
    coloresDisponibles: string | null;
    tallasDisponibles: string | null;
    fechacreacion: Date | string | null;
    fechaultactualizacion: Date | string | null;
    esServicio: number | null;
    productogrupoCodigo: number | null;
    marcaId: number | null;
    productotipoId: number | null;
    ivaporcent: number | null;
    iceporcent: number | null;
    impverdeporcent: number | null;
    otroimpuestoporcent: number | null;
    descmaxporcent: number | null;
    codbarras1: string | null;
    codbarras2: string | null;
    codbarras3: string | null;
    codbarras4: string | null;
    codbarras5: string | null;
    codbarras6: string | null;
    descuentomaximoporcentunico: number | null;
    pvppromo: number | null;
    finpvppromo: string | null;
    fechafinpromo: Date | string | null;
    ajuste1: number | null;
    ajuste2: number | null;
    ajuste2cuantos: number | null;
    esPromo: number | null;
    esPromo2: number | null;
    cuantosPromo: number | null;
    costoInventario: number | null;
    especificaciones: number | null;
    estado: number | null;
    tipoCuenta: string | null;
    tipoCuentaActivo: string | null;
    tipoCuentaVentas: string | null;
    vidaUtil: number | null;
    prodFechaCaducidad: Date | string | null;
    esSuperproducto: number | null;
    img: string | null;
    unidadMedida: string | null;
    valorMedida: number | null;
    idSubgrupo: number | null;
    proteinas: number | null;
    calorias: number | null;
    tipoTela: string | null;
    color: string | null;
    ubicacion: string | null;
    peso: string | null;
    rendimiento: string | null;
    composicion: string | null;
    contraindicaciones: string | null;
    consumo: string | null;
    nombreMedida: string | null;
    largo: number | null;
    ancho: number | null;
    pieza: number | null;
    pvpServicio: number | null;
    servicioEstudiante: number | null;
    imgWeb: string | null;
    tiempo: number | null;
    vistaWeb: number | null;
    destacado: number | null;
    fechaDestacado: Date | string | null;
    imagenUno: string | null;
    imagenDos: string | null;
    imagenTres: string | null;
    imagenCuatro: string | null;
    masVendido: number | null;
    origen: number | null;
    cantBulto: string | null;
    formaFarmaceutica: string | null;
    concentracion: string | null;
    nivelPrescripcion: string | null;
    guarniciones: number | null;
    idBodegas: string | null;
    documento: string | null;
    fechaCompra: Date | string | null;
    imgRepositorio: string | null;
    proveedor: string | null;
    descuentoActivo: number | null;
    seccionPrint: number | null;
    tipoCuentaCosto: string | null;
    presentacion: string | null;
    descuentoId: number | null;
  }